requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
httpie==2.3.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pygments==2.7.3
PySocks==1.7.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
